implemented SupportPage on Selenide side. minor refactoring.
