+++ src/test/java/tests/fixtures/SeleniumFixture.java
@@ -1,6 +1,6 @@
 package tests.fixtures;
 
-import com.alexanderbakhin.site.MyWebSite;
+import com.alexanderbakhin.site.selenium.MyWebSite;
 import driver.WebDriverInstance;
 import lombok.extern.log4j.Log4j;
 import org.junit.AfterClass;
